Spartox v0.0.1.1
Inventory UI added and items can be added by pressing key R and removed by pressing key T.

+++ Source/Spartox/Core/Items/Structs/FItemStruct.h
@@ -32,7 +32,7 @@ struct FItemStruct : public FTableRowBase
 		TEnumAsByte<EItemRarity> Rarity;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
-		uint8 Quantity;
+		int32 Quantity;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
 		bool bStackable;
